[feat] create dependency container
- 디펜던시 컨테이너 만들어서 알고리즘이랑 애니메이션 생성담당
